Make sure blocks are always ordered
The `for_each_concurrent` adapter does not guarnatee any ordering.
Specifically, once the writing of the block is offloaded to the async_fs
thread pool, the blocks can be written in any order, and the tasks can
be resumed in any order. As such, it is possible for some chunks to be
in a different order than they were uploaded, causing file corruption.

Signed-off-by: Lee Smet <lee.smet@hotmail.com>